+++ cache/q.go
@@ -3,7 +3,7 @@ package cache
 import (
 	"fmt"
 
-	"github.com/patrickmn/go-cache"
+	cache "github.com/patrickmn/go-cache"
 	"gitlab.com/tokend/horizon/db2/core"
 	"gitlab.com/tokend/horizon/db2/history"
 )