+++ db2/core2/balance_q.go
@@ -39,7 +39,7 @@ func (q BalancesQ) FilterByAccount(accountAddress string) BalancesQ {
 }
 
 // FilterByAsset - returns new Q with added filter for asset
-func (q BalancesQ) FilterByAsset (asset string) BalancesQ {
+func (q BalancesQ) FilterByAsset(asset string) BalancesQ {
 	q.selector = q.selector.Where("balances.asset = ?", asset)
 	return q
 }