+++ db2/history/reviewable_request_details.go
@@ -3,12 +3,13 @@ package history
 import (
 	"time"
 
+	"gitlab.com/tokend/regources"
+
 	"database/sql/driver"
 
 	"gitlab.com/distributed_lab/logan/v3/errors"
 	"gitlab.com/tokend/go/xdr"
 	"gitlab.com/tokend/horizon/db2"
-	"gitlab.com/tokend/regources"
 )
 
 type ReviewableRequestDetails struct {