+++ web_v2/handlers/get_create_asset_requests.go
@@ -5,6 +5,7 @@ import (
 
 	"gitlab.com/distributed_lab/logan/v3/errors"
 	"gitlab.com/tokend/horizon/web_v2/resources"
+	regources "gitlab.com/tokend/regources/generated"
 
 	"gitlab.com/tokend/horizon/db2/core2"
 
@@ -16,7 +17,6 @@ import (
 	"gitlab.com/tokend/horizon/db2/history2"
 	"gitlab.com/tokend/horizon/web_v2/ctx"
 	"gitlab.com/tokend/horizon/web_v2/requests"
-	"gitlab.com/tokend/regources/v2"
 )
 
 func GetCreateAssetRequests(w http.ResponseWriter, r *http.Request) {