+++ web_v2/handlers/get_order_book.go
@@ -1,6 +1,8 @@
 package handlers
 
 import (
+	"net/http"
+
 	"gitlab.com/distributed_lab/ape"
 	"gitlab.com/distributed_lab/ape/problems"
 	"gitlab.com/distributed_lab/logan/v3"
@@ -10,8 +12,7 @@ import (
 	"gitlab.com/tokend/horizon/web_v2/ctx"
 	"gitlab.com/tokend/horizon/web_v2/requests"
 	"gitlab.com/tokend/horizon/web_v2/resources"
-	"gitlab.com/tokend/regources/v2"
-	"net/http"
+	regources "gitlab.com/tokend/regources/generated"
 )
 
 // GetOrderBook - processes request to get order book entries
@@ -56,7 +57,7 @@ type getOrderBookHandler struct {
 const secondaryMarketOrderBookID = 0
 
 // GetOrderBook returns offer with related resources
-func (h *getOrderBookHandler) GetOrderBook(request *requests.GetOrderBook) (*regources.OrderBookEntriesResponse, error) {
+func (h *getOrderBookHandler) GetOrderBook(request *requests.GetOrderBook) (*regources.OrderBookEntrysResponse, error) {
 	if request.ID != secondaryMarketOrderBookID {
 		sale, err := h.SalesQ.GetByID(request.ID)
 		if err != nil {
@@ -95,7 +96,7 @@ func (h *getOrderBookHandler) GetOrderBook(request *requests.GetOrderBook) (*reg
 		return nil, errors.Wrap(err, "Failed to get offer list")
 	}
 
-	response := &regources.OrderBookEntriesResponse{
+	response := &regources.OrderBookEntrysResponse{
 		Data:  make([]regources.OrderBookEntry, 0, len(coreOrderBookEntries)),
 		Links: request.GetOffsetLinks(*request.PageParams),
 	}